Solution to Input and Print a Double

+++ Input and Print a Double@@6788d5012db25efec946f942/code.c
@@ -5,7 +5,7 @@ char* welcome() {
 }
 
 int main() {
-    float numb;
+    double numb;
     scanf("%f",&numb);
     printf("You entered: %.4f",numb);
     return 0;
